get rid of some compiler warnings

+++ fmt/fmt_longlong.c
@@ -3,7 +3,7 @@
 size_t fmt_longlong(char *dest,signed long long int i) {
   if (i<0) {
     if (dest) *dest++='-';
-    return fmt_ulonglong(dest,-i)+1;
+    return fmt_ulonglong(dest,(unsigned long long)-i)+1;
   } else
-    return fmt_ulonglong(dest,i);
+    return fmt_ulonglong(dest,(unsigned long long)i);
 }
